v1.1.3 : Travis functional testing & minor changes

+++ deployment/build_last_package.sh
@@ -5,21 +5,51 @@ COL_RESET=$ESC_SEQ"39;49;00m"
 COL_RED=$ESC_SEQ"31;01m"
 COL_GREEN=$ESC_SEQ"32;01m"
 COL_YELLOW=$ESC_SEQ"33;01m"
-COL_BLUE=$ESC_SEQ"34;01m"
-COL_MAGENTA=$ESC_SEQ"35;01m"
-COL_CYAN=$ESC_SEQ"36;01m"
 
 echo -e $COL_YELLOW"Deleting old package ... "$COL_RESET
 rm -f cameradar_*_${1:-"Release"}_Linux.tar.gz
 echo -e $COL_GREEN"OK!"$COL_RESET
 
 echo -e $COL_YELLOW"Creating package ... "$COL_RESET
-  cd ..
-  mkdir build
-  cd build
-  rm -f cameradar_*_${1:-"Release"}_Linux.tar.gz
-  cmake .. -DCMAKE_BUILD_TYPE=${1:-"Release"}
-  make package
-  cp cameradar_*_${1:-"Release"}_Linux.tar.gz ../deployment
-  cd ../deployment
+
+cd ..
+ret=$?
+if [ "$ret" -ne "0" ]; then
+  echo -e $COL_RED"KO!"$COL_RESET;
+  exit 1;
+fi
+
+mkdir build
+
+cd build
+ret=$?
+if [ "$ret" -ne "0" ]; then
+  echo -e $COL_RED"KO!"$COL_RESET;
+  exit 1;
+fi
+
+rm -f cameradar_*_${1:-"Release"}_Linux.tar.gz
+
+cmake .. -DCMAKE_BUILD_TYPE=${1:-"Release"}
+ret=$?
+if [ "$ret" -ne "0" ]; then
+  echo -e $COL_RED"KO!"$COL_RESET;
+  exit 1;
+fi
+
+make package
+ret=$?
+if [ "$ret" -ne "0" ]; then
+  echo -e $COL_RED"KO!"$COL_RESET;
+  exit 1;
+fi
+
+cp cameradar_*_${1:-"Release"}_Linux.tar.gz ../deployment
+
+cd ../deployment
+ret=$?
+if [ "$ret" -ne "0" ]; then
+  echo -e $COL_RED"KO!"$COL_RESET;
+  exit 1;
+fi
 echo -e $COL_GREEN"OK!"$COL_RESET
